Move logic of document kind resolving into separate class
